requirements.txt
# Kermartin 3.0 - Dependências de Produção

# Django Core
Django>=4.2.0,<5.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
django-cors-headers>=4.3.0

# Banco de Dados
psycopg2-binary>=2.9.0  # PostgreSQL
dj-database-url>=2.1.0

# Servidor Web
gunicorn>=21.2.0
whitenoise>=6.6.0

# IA e Processamento
openai>=1.3.0
PyMuPDF>=1.23.0
pdfplumber>=0.10.0

# Cache e Performance
redis>=5.0.0
django-redis>=5.4.0

# Configuração
python-decouple>=3.8
python-dotenv>=1.0.0

# Monitoramento e Logs
sentry-sdk[django]>=1.38.0

# Utilitários
requests>=2.31.0
Pillow>=10.0.0
python-dateutil>=2.8.0
beautifulsoup4>=4.12.2
serpapi>=0.1.5

# Desenvolvimento e Testes
pytest>=7.4.0
pytest-django>=4.7.0

# Segurança
cryptography>=41.0.0
